mgr/cephadm: Move remove osd code to osd.py
