+++ src/pybind/mgr/tox.ini
@@ -14,7 +14,6 @@ deps =
     mypy
 commands = mypy --config-file=../../mypy.ini \
            cephadm/module.py \
-           cephadm/_utils.py \
            mgr_module.py \
            dashboard/module.py \
            mgr_util.py \
